+++ app/src/main/java/projects/feng/gary/matrixtools/Fraction.kt
@@ -1,9 +1,11 @@
 package projects.feng.gary.matrixtools
 
+import java.math.BigInteger
+
 
 class Fraction {
-    private val numerator: Int
-    private val denominator: Int
+    private val numerator: BigInteger
+    private val denominator: BigInteger
 
     companion object {
         val zero = Fraction(0);
@@ -13,16 +15,16 @@ class Fraction {
 
     //---------------------------CONSTRUCTORS-----------------------------------------------------//
 
-    constructor(numer: Int, denom: Int) {
-        val gcd = if (denom < 0) -gcd(numer, denom) else gcd(numer, denom)
+    constructor(numer: BigInteger, denom: BigInteger) {
+        val gcd = if (denom < 0.toBigInteger()) -gcd(numer, denom) else gcd(numer, denom)
 
         numerator = numer / gcd
         denominator = denom / gcd
     }
 
     constructor(integer: Int) {
-        numerator = integer
-        denominator = 1
+        numerator = integer.toBigInteger()
+        denominator = 1.toBigInteger()
     }
 
     constructor(str: String) {
@@ -30,16 +32,16 @@ class Fraction {
 
         when {
             stringRep.equals("") -> {
-                numerator = 0
-                denominator = 1
+                numerator = 0.toBigInteger()
+                denominator = 1.toBigInteger()
             }
             stringRep.contains('/') -> {
                 val fracParts = stringRep.split("/".toRegex())
 
-                val numer = fracParts.get(0).toInt()
-                val denom = fracParts.get(1).toInt()
+                val numer = fracParts.get(0).toBigInteger()
+                val denom = fracParts.get(1).toBigInteger()
 
-                val gcd = if (denom < 0) -gcd(numer, denom) else gcd(numer, denom)
+                val gcd = if (denom < 0.toBigInteger()) -gcd(numer, denom) else gcd(numer, denom)
 
                 numerator = numer / gcd
                 denominator = denom / gcd
@@ -49,22 +51,22 @@ class Fraction {
                 val numDecPoints = decParts[1].length
 
                 // reduce fractional component before adding on the integer component
-                var fracPartNum = if (numDecPoints == 0) 0 else decParts[1].toInt()
+                var fracPartNum = if (numDecPoints == 0) 0.toBigInteger() else decParts[1].toBigInteger()
                 var fracPartDenom = powerOfTen(numDecPoints)
 
                 val gcd = gcd(fracPartNum, fracPartDenom)
                 fracPartNum /= gcd
                 fracPartDenom /= gcd
 
-                val intPart = decParts[0].toInt() * fracPartDenom
+                val intPart = decParts[0].toBigInteger() * fracPartDenom
 
                 numerator = (fracPartNum + intPart) *
-                        if (decParts[0].toInt() == 0 && decParts[0].contains('-')) -1 else 1
+                        if (decParts[0].toInt() == 0 && decParts[0].contains('-')) -1.toBigInteger() else 1.toBigInteger()
                 denominator = fracPartDenom
             }
             else -> {
-                numerator = stringRep.toInt()
-                denominator = 1
+                numerator = stringRep.toBigInteger()
+                denominator = 1.toBigInteger()
             }
         }
     }
@@ -112,7 +114,7 @@ class Fraction {
     }
 
     override fun toString(): String {
-        return if (this.denominator == 1)
+        return if (this.denominator == 1.toBigInteger())
             this.numerator.toString()
         else
             this.numerator.toString() + "/" + this.denominator.toString()
@@ -121,15 +123,15 @@ class Fraction {
 
     //---------------------------HELPERS----------------------------------------------------------//
 
-    private fun gcd(x: Int, y: Int): Int = if (y == 0) x.abs() else gcd(y, x.rem(y))
+    private fun gcd(x: BigInteger, y: BigInteger): BigInteger = if (y == 0.toBigInteger()) x.abs() else gcd(y, x.rem(y))
 
-    private fun powerOfTen(n: Int): Int {
+    private fun powerOfTen(n: Int): BigInteger {
         var result = 1
 
         for (i in 1..n) {
             result *= 10
         }
 
-        return result
+        return result.toBigInteger()
     }
 }
